Rename game_card.php -> GameCard.php, for consistency, and remove unused import.
