+++ site/components/card.php
@@ -1,5 +1,4 @@
 <?php 
-require_once 'models/game_card.php';
 require_once 'includes/functions.php';
 
 class Card
